requirements.txt
imageio==2.34.1
Flask==3.0.3
joblib==1.4.2
numpy==1.26.4
tensorflow==2.17.0
pillow==10.2.0
pandas==2.2.2